requirements.txt
rasa-sdk==3.6.2
python-dotenv==1.0.1
pyodbc==5.1.0